Bug #3651 - revert to 32 bit hashing
Revert "Bug #3381 - make rehash tool 64-bit safe"

This reverts commit e4712a63cc650bc064e456fcbe608cb74525c24b.

+++ lib/util.c
@@ -269,7 +269,7 @@ int dir_hash_c(const char *name, int full)
 
     if (full) {
 	unsigned char *pt;
-	unsigned long long int n;
+	unsigned int n;
 	enum {
 	    DIR_X = 3,
 	    DIR_Y = 5,
